# documentation/api/OPENAPI-3.0.yaml
openapi: 3.0.0
info:
  title: Sugar-Daddy Platform - Comprehensive API
  description: |
    完整的 Sugar-Daddy 平台 API 文檔
    包含所有 81 個端點的詳細說明
    
    ### 系統組成
    - 🔐 Auth Service (11 個端點)
    - 🎬 Content-Streaming Service (11 個端點)  
    - 💳 Payment Service (19 個端點)
    - 🎁 Subscription Service (10 個端點)
    - 🤖 Recommendation Service (13 個端點)
    - 🌉 API Gateway (1 個端點)
    
    **總計: 81 個 API 端點**
  version: 1.0.0
  contact:
    name: Sugar-Daddy Backend Team
    email: backend@sugardaddy.com
  license:
    name: Proprietary
    url: https://sugardaddy.com/license
servers:
  - url: http://localhost:3000
    description: API Gateway (本地開發)
    variables:
      port:
        default: '3000'
  - url: https://api.sugardaddy.com/v1
    description: 生產環境
  - url: https://staging-api.sugardaddy.com/v1
    description: 預發佈環境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token 認證
        登錄後獲得 token，在 Authorization header 中使用
        格式: Authorization: Bearer <token>
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Bad Request
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - statusCode
        - message

    # ==================== Auth Service Schemas ====================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        username:
          type: string
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: admin
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        resource:
          type: string
          example: users
        action:
          type: string
          example: create
        description:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          example: 99.99
        currency:
          type: string
          example: USD
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        paymentMethod:
          type: string
          enum: [credit_card, paypal, stripe, alipay, wechat]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, paused, cancelled, expired]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        autoRenew:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
            errors:
              - field: token
                message: Invalid or expired token
    NotFoundError:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not Found
            errors:
              - message: Resource not found
    BadRequestError:
      description: 請求格式錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/auth/users:
    get:
      tags:
        - Auth
      summary: 獲取用戶列表
      description: 獲取所有用戶列表（需要管理員權限）
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        '200':
          description: 用戶列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          total:
                            type: integer
                          limit:
                            type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Auth
      summary: 創建用戶
      description: 創建新用戶（需要管理員權限）
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
              required:
                - username
                - email
      responses:
        '201':
          description: 用戶創建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/users/{userId}:
    get:
      tags:
        - Auth
      summary: 獲取用戶詳細信息
      description: 獲取指定用戶的詳細信息
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶詳細信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Auth
      summary: 刪除用戶
      description: 刪除指定用戶（需要管理員權限）
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶刪除成功
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/users/{userId}/activate:
    post:
      tags:
        - Auth
      summary: 激活用戶
      description: 激活已禁用的用戶賬戶
      operationId: activateUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶激活成功
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/users/{userId}/deactivate:
    post:
      tags:
        - Auth
      summary: 禁用用戶
      description: 禁用用戶賬戶
      operationId: deactivateUser
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶禁用成功
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/auth/roles:
    get:
      tags:
        - Auth
      summary: 獲取角色列表
      description: 獲取系統中所有角色
      operationId: listRoles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Auth
      summary: 創建角色
      description: 創建新角色（需要管理員權限）
      operationId: createRole
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: moderator
                description:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
              required:
                - name
      responses:
        '201':
          description: 角色創建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/roles/{roleId}:
    get:
      tags:
        - Auth
      summary: 獲取角色詳細信息
      description: 獲取指定角色的詳細信息
      operationId: getRoleById
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 角色詳細信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Role'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    patch:
      tags:
        - Auth
      summary: 更新角色
      description: 更新角色信息（需要管理員權限）
      operationId: updateRole
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: 角色更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/roles/{roleId}/permissions:
    get:
      tags:
        - Auth
      summary: 獲取角色權限
      description: 獲取指定角色的所有權限
      operationId: getRolePermissions
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 角色權限列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Auth
      summary: 添加角色權限
      description: 為角色添加權限（需要管理員權限）
      operationId: addRolePermission
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissionId:
                  type: string
                  format: uuid
              required:
                - permissionId
      responses:
        '200':
          description: 權限添加成功
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==================== Permission Management ====================
  /api/auth/permissions:
    get:
      tags:
        - Auth
      summary: 獲取權限列表
      description: 獲取系統中所有可用的權限
      operationId: listPermissions
      security:
        - BearerAuth: []
      parameters:
        - name: resource
          in: query
          schema:
            type: string
          description: 按資源篩選
      responses:
        '200':
          description: 權限列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==================== Content-Streaming Service ====================
...
  /api/payments/user/{userId}:
    get:
      tags:
        - Payment
      summary: 獲取特定用戶的支付記錄
      description: 獲取指定用戶的所有支付記錄（需要管理員權限）
      operationId: getUserPayments
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶支付記錄
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/subscriptions/user/{userId}:
    get:
      tags:
        - Subscription
      summary: 獲取特定用戶的訂閱
      description: 獲取指定用戶的所有訂閱（需要管理員權限）
      operationId: getUserSubscriptions
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 用戶訂閱列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subscription'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==================== Recommendation Service ====================
...
security:
  - BearerAuth: []